Create K-Factorisation.cpp
a program for solving a problem that involves finding a sequence of values (factors) that, when multiplied together, result in a given number N. The goal is to find the sequence with the smallest values in ascending order. The problem seems to be constrained by certain rules, such as the number of factors (K) and specific values for the factors.